some clean and respect of c++98

+++ includes/Response.hpp
@@ -65,7 +65,7 @@ private:
 	void			setContentType(std::string contentType);
 
 public:
-	Response(Request *req, int socket);
+	Response(Request *req, long socket);
 	Response(Response const & src);
 	~Response(void);
 	Response & operator=(Response const & rhs);
